requirements.txt
aiogram==3.22.0
aiohttp>=3.9.0
redis>=5.0.0
supabase>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.6.0